+++ tweed5/default-extensions/src/main/java/de/siphalor/tweed5/defaultextensions/pather/api/PatherExtension.java
@@ -1,10 +1,11 @@
 package de.siphalor.tweed5.defaultextensions.pather.api;
 
 import de.siphalor.tweed5.core.api.extension.TweedExtension;
+import de.siphalor.tweed5.defaultextensions.pather.impl.PatherExtensionImpl;
 import de.siphalor.tweed5.serde.extension.api.TweedReadContext;
 import de.siphalor.tweed5.serde.extension.api.TweedWriteContext;
-import de.siphalor.tweed5.defaultextensions.pather.impl.PatherExtensionImpl;
 
+@Deprecated
 public interface PatherExtension extends TweedExtension {
 	Class<? extends PatherExtension> DEFAULT = PatherExtensionImpl.class;
 	String EXTENSION_ID = "pather";