+++ tweed5/default-extensions/src/main/java/de/siphalor/tweed5/defaultextensions/readfallback/impl/ReadFallbackExtensionImpl.java
@@ -6,7 +6,6 @@ import de.siphalor.tweed5.core.api.middleware.Middleware;
 import de.siphalor.tweed5.serde.extension.api.TweedEntryReader;
 import de.siphalor.tweed5.serde.extension.api.extension.ReadWriteExtensionSetupContext;
 import de.siphalor.tweed5.serde.extension.api.extension.ReadWriteRelatedExtension;
-import de.siphalor.tweed5.defaultextensions.pather.api.PatherExtension;
 import de.siphalor.tweed5.defaultextensions.presets.api.PresetsExtension;
 import de.siphalor.tweed5.defaultextensions.readfallback.api.ReadFallbackExtension;
 import de.siphalor.tweed5.defaultextensions.validation.api.ValidationExtension;
@@ -39,7 +38,7 @@ public class ReadFallbackExtensionImpl implements ReadFallbackExtension, ReadWri
 
 			@Override
 			public Set<String> mustComeBefore() {
-				return Collections.singleton(PatherExtension.EXTENSION_ID);
+				return Collections.emptySet();
 			}
 
 			@Override