+++ tweed5/serde-extension/src/main/java/de/siphalor/tweed5/serde/extension/api/TweedReadContext.java
@@ -3,6 +3,7 @@ package de.siphalor.tweed5.serde.extension.api;
 import de.siphalor.tweed5.core.api.entry.ConfigEntry;
 import de.siphalor.tweed5.core.api.entry.SubEntryKey;
 import de.siphalor.tweed5.patchwork.api.Patchwork;
+import de.siphalor.tweed5.serde.extension.api.path.EntryPath;
 import de.siphalor.tweed5.serde.extension.api.read.result.TweedReadResult;
 import de.siphalor.tweed5.serde_api.api.TweedDataReader;
 import org.jspecify.annotations.Nullable;
@@ -10,6 +11,8 @@ import org.jspecify.annotations.Nullable;
 public interface TweedReadContext {
 	ReadWriteExtension readWriteExtension();
 	Patchwork extensionsData();
+	EntryPath currentEntryPath();
+	EntryPath currentValuePath();
 
 	<T extends @Nullable Object, C extends ConfigEntry<T>> TweedReadResult<T> readSubEntry(
 			TweedDataReader reader, C entry, SubEntryKey key