+++ tweed5/serde-extension/src/main/java/de/siphalor/tweed5/serde/extension/api/TweedWriteContext.java
@@ -3,12 +3,15 @@ package de.siphalor.tweed5.serde.extension.api;
 import de.siphalor.tweed5.core.api.entry.ConfigEntry;
 import de.siphalor.tweed5.core.api.entry.SubEntryKey;
 import de.siphalor.tweed5.patchwork.api.Patchwork;
+import de.siphalor.tweed5.serde.extension.api.path.EntryPath;
 import de.siphalor.tweed5.serde_api.api.TweedDataVisitor;
 import de.siphalor.tweed5.serde_api.api.TweedDataWriteException;
 import org.jspecify.annotations.Nullable;
 
 public interface TweedWriteContext {
 	Patchwork extensionsData();
+	EntryPath currentEntryPath();
+	EntryPath currentValuePath();
 
 	<T extends @Nullable Object, C extends ConfigEntry<T>> void writeSubEntry(
 			TweedDataVisitor writer, C entry, SubEntryKey key, @Nullable T value