+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `core`: Fixed return type of `MutableStructuredConfigEntry#apply` to correctly be itself.
 - `minecraft-fabric-helper`: Fixed missing new line in error log message.
+- `serde-extension`, `default-extensions`: Integrate path tracking into `ReadWriteExtension`, deprecate the existing
+  `PatherExtension`
 - `type-utls`: Fixed missing `@Nullable` annotation on `ActualType#getAnnotation`.
 - `weaver-pojo`: Fixed `StringMapPojoWeaver` weaving entries without `@StringMapWeaving` annotation.